A8 Demo Video BEFORE presentation
NOTE: this is the demo video before our presentation since we could not record a new version that had the bug fixes and minor text fixes after our presentation. Hence why it may look a bit different then what the turn in code makes it look like. All the features are still there and have not changed however. Quality isn't the greatest due to needing <25MB to upload